Moves check_and_parse_page tests to separate file
